+++ Jellyfin.Plugin.JellyfinHelper.Tests/PluginPages/ConfigPageHtmlTests.cs
@@ -1,7 +1,7 @@
 using System.Text.RegularExpressions;
 using Xunit;
 
-namespace Jellyfin.Plugin.JellyfinHelper.Tests;
+namespace Jellyfin.Plugin.JellyfinHelper.Tests.PluginPages;
 
 /// <summary>
 /// General page structure and README quality tests.