+++ Jellyfin.Plugin.JellyfinHelper.Tests/ScheduledTasks/HelperCleanupTaskTests.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Xunit;
 
-namespace Jellyfin.Plugin.JellyfinHelper.Tests;
+namespace Jellyfin.Plugin.JellyfinHelper.Tests.ScheduledTasks;
 
 /// <summary>
 /// Tests for <see cref="HelperCleanupTask"/>, the master orchestration task.