+++ Jellyfin.Plugin.JellyfinHelper.Tests/ScheduledTasks/RepairStrmFilesTaskTests.cs
@@ -5,7 +5,7 @@ using MediaBrowser.Controller.Library;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace Jellyfin.Plugin.JellyfinHelper.Tests;
+namespace Jellyfin.Plugin.JellyfinHelper.Tests.ScheduledTasks;
 
 /// <summary>
 /// Tests for <see cref="RepairStrmFilesTask"/>.