+++ Jellyfin.Plugin.JellyfinHelper.Tests/Services/I18nServiceTests.cs
@@ -5,7 +5,7 @@ using Jellyfin.Plugin.JellyfinHelper.Configuration;
 using Jellyfin.Plugin.JellyfinHelper.Services;
 using Xunit;
 
-namespace Jellyfin.Plugin.JellyfinHelper.Tests;
+namespace Jellyfin.Plugin.JellyfinHelper.Tests.Services;
 
 public class I18nServiceTests
 {