+++ Jellyfin.Plugin.JellyfinHelper.Tests/Services/TrashServiceTests.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Xunit;
 
-namespace Jellyfin.Plugin.JellyfinHelper.Tests;
+namespace Jellyfin.Plugin.JellyfinHelper.Tests.Services;
 
 public class TrashServiceTests : IDisposable
 {